ADding maya2017/2016 changes to the centos7 build

+++ linux/build-centos6_maya2017.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "Copy local root certificates for corporate networks"
-cp -u /etc/pki/ca-trust/source/anchors/* cert/
+[ -e /etc/pki/ca-trust/source/anchors ] && cp -u /etc/pki/ca-trust/source/anchors/* cert/
 
 echo "Build base: base centos packages and gcc"
 docker build -t "usd-docker/base:centos6-usd-0.7" -f centos6/base/Dockerfile .
